Move shareable code to WasmObjectWriter.cs
