+++ src/coreclr/tools/Common/Compiler/ObjectWriter/WebCilObjectWriter.cs
@@ -17,7 +17,6 @@ using Internal.Text;
 using Internal.TypeSystem;
 using Microsoft.NET.WebAssembly.Webcil;
 using CodeDataLayout = CodeDataLayoutMode.CodeDataLayout;
-using ObjectData = ILCompiler.DependencyAnalysis.ObjectNode.ObjectData;
 
 namespace ILCompiler.ObjectWriter
 {
@@ -38,33 +37,13 @@ namespace ILCompiler.ObjectWriter
         }
     }
 
-    internal static class WasmObjectNodeSection
-    {
-        // TODO-WASM: Consider alignment needs for data sections
-        public static readonly ObjectNodeSection DataCountSection = new ObjectNodeSection("wasm.datacount", SectionType.ReadOnly, needsAlign: false);
-        public static readonly ObjectNodeSection CombinedDataSection = new ObjectNodeSection("wasm.alldata", SectionType.Writeable, needsAlign: false);
-        public static readonly ObjectNodeSection FunctionSection = new ObjectNodeSection("wasm.function", SectionType.ReadOnly, needsAlign: false);
-        public static readonly ObjectNodeSection ExportSection = new ObjectNodeSection("wasm.export", SectionType.ReadOnly, needsAlign: false);
-        public static readonly ObjectNodeSection ElementSection = new ObjectNodeSection("wasm.element", SectionType.ReadOnly, needsAlign: false);
-        public static readonly ObjectNodeSection MemorySection = new ObjectNodeSection("wasm.memory", SectionType.ReadOnly, needsAlign: false);
-        public static readonly ObjectNodeSection TableSection = new ObjectNodeSection("wasm.table", SectionType.ReadOnly, needsAlign: false);
-        public static readonly ObjectNodeSection ImportSection = new ObjectNodeSection("wasm.import", SectionType.ReadOnly, needsAlign: false);
-        public static readonly ObjectNodeSection GlobalSection = new ObjectNodeSection("wasm.global", SectionType.ReadOnly, needsAlign: false);
-    }
-
-    internal abstract partial class WasmObjectWriter : ObjectWriter
-    {
-        protected WasmObjectWriter(NodeFactory factory, ObjectWritingOptions options, OutputInfoBuilder outputInfoBuilder)
-            : base(factory, options, outputInfoBuilder)
-        {
-        }
-    }
-
     /// <summary>
     /// WebCIL object file format writer.
     /// </summary>
     internal sealed class WebCilObjectWriter : WasmObjectWriter
     {
+        public const int WebcilSectionAlignment = 16;
+
         protected override CodeDataLayout LayoutMode => CodeDataLayout.Separate;
 
         // We use 2 Wasm data segments for webcil,
@@ -75,17 +54,6 @@ namespace ILCompiler.ObjectWriter
             : base(factory, options, outputInfoBuilder)
         {
         }
-
-        private void EmitWasmHeader(Stream outputFileStream)
-        {
-            outputFileStream.Write("\0asm"u8);
-            outputFileStream.Write([0x1, 0x0, 0x0, 0x0]);
-        }
-
-        /// <summary>
-        /// The number of methods in the Function section.
-        /// </summary>
-        private int MethodCount => _wasmSymbolManager.GetDefinitionCount(WasmIndexSpace.Function);
 
         private Dictionary<SortableDependencyNode.ObjectNodeOrder, Utf8String> _wellKnownSymbols = new();
         private protected override void RecordWellKnownSymbol(Utf8String currentSymbolName, SortableDependencyNode.ObjectNodeOrder classCode)
@@ -100,180 +68,16 @@ namespace ILCompiler.ObjectWriter
             }
         }
 
-        private protected override void RecordMethodSignature(WasmTypeNode signature)
+        private protected override SectionDataEmitter CreateDataSection(
+            ObjectNodeSection section,
+            int sectionIndex,
+            Stream sectionStream)
         {
-            var mangledNameBuilder = new Utf8StringBuilder();
-            signature.AppendMangledName(_nodeFactory.NameMangler, mangledNameBuilder);
-            Utf8String mangledName = mangledNameBuilder.ToUtf8String();
-            // Record the signature's wasm type index in the shared symbol table. The signature bytes
-            // are emitted by the node's own data; here we only assign its index.
-            _wasmSymbolManager.AddDefinition(mangledName, WasmIndexSpace.Type);
-        }
-
-        private protected override void RecordMethodDeclaration(INodeWithTypeSignature node)
-        {
-            WasmLowering.LoweringFlags flags = WasmLowering.LoweringFlags.None;
-            if (node.HasGenericContextArg)
-            {
-                flags |= WasmLowering.LoweringFlags.HasGenericContextArg;
-            }
-            if (node.IsAsyncCall)
-            {
-                flags |= WasmLowering.LoweringFlags.IsAsyncCall;
-            }
-            if (node.IsUnmanagedCallersOnly)
-            {
-                flags |= WasmLowering.LoweringFlags.IsUnmanagedCallersOnly;
-            }
-            WriteSignatureIndexForFunction(node.Signature, flags, node);
-            RegisterFunctionSymbol(new Utf8String(node.GetMangledName(_nodeFactory.NameMangler)));
-            if (node is INodeWithFunclets nodeWithFunclets)
-            {
-                RecordFunclets(nodeWithFunclets);
-            }
-        }
-
-        private void RecordFunclets(INodeWithFunclets nodeWithFunclets)
-        {
-            FuncletKind[] funcletKinds = nodeWithFunclets.GetFuncletKinds();
-            if (funcletKinds.Length < 1)
-            {
-                return;
-            }
-
-            WasmValueType pointerType = _nodeFactory.Target.PointerSize == 8 ? WasmValueType.I64 : WasmValueType.I32;
-            string mangledNodeName = nodeWithFunclets.GetMangledName(_nodeFactory.NameMangler);
-
-            for (int i = 0; i < funcletKinds.Length; i++)
-            {
-                WasmFuncType funcletSignature = GetFuncletType(funcletKinds[i], pointerType);
-                RegisterFunctionSymbol(new Utf8String($"{mangledNodeName}_funclet_{i}"));
-                RegisterStubIndexAndSignature(funcletSignature);
-            }
-        }
-
-        private WasmFuncType GetFuncletType(FuncletKind funcletKind, WasmValueType pointerType)
-        {
-            return funcletKind switch
-            {
-                FuncletKind.CatchOrFilterHandler or FuncletKind.Filter => new WasmFuncType(
-                    new([pointerType, pointerType, pointerType]), new([pointerType])), // (FP, SP, EXN) -> RESULT
-                _ => new WasmFuncType(new([pointerType, pointerType]), new([])), // (FP, SP) -> void
-            };
-        }
-
-        private void WriteFunctionEntry(int signatureIndex)
-        {
-            WasmFunctionSection section = GetOrCreateSection<WasmFunctionSection>(
-                WasmObjectNodeSection.FunctionSection,
-                out SectionWriter writer);
-            section.WriteEntry(writer, signatureIndex);
-        }
-
-        private void WriteSignatureIndexForFunction(MethodSignature managedSignature, WasmLowering.LoweringFlags flags, ISymbolNode node)
-        {
-            WasmFuncType signature = WasmLowering.GetSignature(managedSignature, flags).FuncType;
-            Utf8String key = signature.GetMangledName(_nodeFactory.NameMangler);
-            if (!_wasmSymbolManager.TryGetSymbol(key, out WasmSymbol signatureSymbol))
-            {
-                throw new InvalidOperationException($"Signature index of {key} not found for function: {node.ToString()}");
-            }
-
-            WriteFunctionEntry(signatureSymbol.Index);
-        }
-
-        /// <summary>
-        /// Adds the given import entry, including its prefix (module/name/kind) and body (external ref).
-        /// </summary>
-        private void WriteImport(WasmImport import)
-        {
-            Utf8String symbolName = new(import.Name);
-            _wasmSymbolManager.AddImport(symbolName, GetIndexSpace(import.Kind), import.Index);
-
-            WasmImportSection section = GetOrCreateSection<WasmImportSection>(
-                WasmObjectNodeSection.ImportSection,
-                out SectionWriter writer);
-            section.WriteEntry(writer, import);
-        }
-
-        private void WriteExport(string name, WasmExportKind kind, int index)
-        {
-            WasmExportSection section = GetOrCreateSection<WasmExportSection>(
-                WasmObjectNodeSection.ExportSection,
-                out SectionWriter writer);
-            section.WriteEntry(writer, new WasmExport(name, kind, index));
-        }
-
-        // Convenience methods for specific export types
-        private void WriteFunctionExport(string name, int functionIndex) =>
-            WriteExport(name, WasmExportKind.Function, functionIndex);
-
-        private void WriteTableExport(string name, int tableIndex) =>
-            WriteExport(name, WasmExportKind.Table, tableIndex);
-
-        private void WriteMemoryExport(string name, int memoryIndex) =>
-            WriteExport(name, WasmExportKind.Memory, memoryIndex);
-
-        private void WriteGlobalExport(string name, int globalIndex) =>
-            WriteExport(name, WasmExportKind.Global, globalIndex);
-
-        private void WriteElementSegment(ReadOnlyMemory<int> functionIndices)
-        {
-            WasmElementSection section = GetOrCreateSection<WasmElementSection>(
-                WasmObjectNodeSection.ElementSection,
-                out SectionWriter writer);
-            section.WriteEntry(writer, functionIndices);
-        }
-
-        private WasmSections _sections = new();
-
-        private TSection GetOrCreateSection<TSection>(ObjectNodeSection section, out SectionWriter writer)
-            where TSection : SectionDataEmitter
-        {
-            writer = base.GetOrCreateSection(section);
-            return _sections.GetSection<TSection>(writer.SectionIndex);
-        }
-
-        private TSection GetOrCreateSection<TSection>(ObjectNodeSection section)
-            where TSection : SectionDataEmitter
-        {
-            var writer = base.GetOrCreateSection(section);
-            return _sections.GetSection<TSection>(writer.SectionIndex);
-        }
-
-        private SectionDataEmitter GetOrCreateSection(ObjectNodeSection section, out SectionWriter writer)
-        {
-            return GetOrCreateSection<SectionDataEmitter>(section, out writer);
-        }
-
-        private new SectionDataEmitter GetOrCreateSection(ObjectNodeSection section)
-        {
-            return GetOrCreateSection<SectionDataEmitter>(section, out _);
-        }
-
-        private Dictionary<ObjectNodeSection, WasmSectionType> _sectionToType = new()
-        {
-            { WasmObjectNodeSection.MemorySection, WasmSectionType.Memory },
-            { WasmObjectNodeSection.FunctionSection, WasmSectionType.Function },
-            { WasmObjectNodeSection.TableSection, WasmSectionType.Table },
-            { WasmObjectNodeSection.ElementSection, WasmSectionType.Element },
-            { WasmObjectNodeSection.ExportSection, WasmSectionType.Export },
-            { WasmObjectNodeSection.ImportSection, WasmSectionType.Import },
-            { WasmObjectNodeSection.GlobalSection, WasmSectionType.Global },
-            { ObjectNodeSection.WasmTypeSection, WasmSectionType.Type },
-            { ObjectNodeSection.WasmCodeSection, WasmSectionType.Code },
-            { WasmObjectNodeSection.DataCountSection, WasmSectionType.DataCount },
-        };
-
-        private WasmSectionType GetWasmSectionType(ObjectNodeSection section)
-        {
-            if (!_sectionToType.ContainsKey(section))
-            {
-                // All other sections map to generic data segments in Wasm
-                // TODO-WASM: Consider making the mapping explicit for every possible node type.
-                return WasmSectionType.Data;
-            }
-            return _sectionToType[section];
+            return new WebcilSection(
+                new Utf8String(section.Name),
+                default(WebcilSectionHeader),
+                sectionStream,
+                sectionIndex);
         }
 
         protected internal override void UpdateSectionAlignment(int sectionIndex, int alignment)
@@ -366,29 +170,6 @@ namespace ILCompiler.ObjectWriter
                 ]
         );
 
-        // This effectively recreates the logic of RecordMethodBody/RecordMethodDeclaration, but for manually inserted stubs that are not
-        // represented by nodes in the dependency graph.
-        // TODO-Wasm: for maintability, we should try and push some of this into the dependency graph when we do more stub generation.
-        private void RegisterStubIndexAndSignature(WasmFuncType signature)
-        {
-            int signatureIndex = RegisterSignature(signature);
-            WriteFunctionEntry(signatureIndex);
-        }
-
-        private void InsertWasmStub(Utf8String name, WasmFunctionBody body)
-        {
-            GetOrCreateSection(ObjectNodeSection.WasmCodeSection, out SectionWriter codeWriter);
-
-            int codeSize = body.EncodeSize();
-            byte[] data = new byte[codeSize];
-            body.Encode(data);
-
-            codeWriter.EmitSymbolDefinition(name);
-            codeWriter.EmitData(data);
-
-            RegisterFunctionSymbol(name);
-            RegisterStubIndexAndSignature(body.Signature);
-        }
         private long ResolveSymbolRVA(WebcilSection[] sections, SymbolDefinition definition)
         {
             for (int i = 0; i < sections.Length; i++)
@@ -402,8 +183,6 @@ namespace ILCompiler.ObjectWriter
 
             return 0;
         }
-
-        public const int WebcilSectionAlignment = 16;
 
         /// <summary>
         /// Assigns VirtualAddresses and related header fields to each webcil section based on the
@@ -500,38 +279,6 @@ namespace ILCompiler.ObjectWriter
             return section;
         }
 
-        private protected override void CreateSection(ObjectNodeSection section, Utf8String comdatName, Utf8String symbolName, int sectionIndex, Stream sectionStream)
-        {
-            WasmSectionType sectionType = GetWasmSectionType(section);
-            SectionDataEmitter wasmSection = null;
-            if (sectionType == WasmSectionType.Data)
-            {
-#if READYTORUN
-                // This is a section which is internally wrapping a Webcil section
-                wasmSection = new WebcilSection(new Utf8String(section.Name), default(WebcilSectionHeader), sectionStream, sectionIndex);
-#else
-                wasmSection = new WasmSection(WasmSectionType.Data, sectionStream, new Utf8String(section.Name), sectionIndex);
-#endif
-            }
-            else
-            {
-                Utf8String sectionName = new(section.Name);
-                wasmSection = sectionType switch
-                {
-                    WasmSectionType.Type or WasmSectionType.Code => new WasmExternallyCountedSection(sectionType, sectionStream, sectionName, sectionIndex),
-                    WasmSectionType.Import => new WasmImportSection(sectionStream, sectionName, sectionIndex),
-                    WasmSectionType.Function => new WasmFunctionSection(sectionStream, sectionName, sectionIndex),
-                    WasmSectionType.Global => new WasmGlobalSection(sectionStream, sectionName, sectionIndex),
-                    WasmSectionType.Export => new WasmExportSection(sectionStream, sectionName, sectionIndex),
-                    WasmSectionType.Element => new WasmElementSection(sectionStream, sectionName, sectionIndex),
-                    _ => new WasmSection(sectionType, sectionStream, sectionName, sectionIndex),
-                };
-            }
-
-            Debug.Assert(_sections.Sections.Count == sectionIndex);
-            _sections.Add(section.Name, sectionIndex, wasmSection);
-        }
-
         private void WriteDataCountSection()
         {
             SectionDataEmitter section = GetOrCreateSection(WasmObjectNodeSection.DataCountSection, out SectionWriter writer);
@@ -550,66 +297,11 @@ namespace ILCompiler.ObjectWriter
             WriteDataCountSection();
         }
 
-        private Dictionary<string, WasmGlobal> _definedGlobals = new();
-
-        // TODO-Wasm: In the future, we may want to consider representing Wasm globals in the dependency graph so that they
-        // can be referenced by other nodes and we can make effective use of them.
-        private void WriteGlobal(string name, WasmValueType valueType, WasmMutabilityType mutability, WasmInstructionGroup initExpr)
-        {
-            Utf8String symbolName = new(name);
-            _wasmSymbolManager.AddDefinition(symbolName, WasmIndexSpace.Global);
-            int index = _wasmSymbolManager.GetSymbol(symbolName).Index;
-            WasmGlobal global = new WasmGlobal(
-                index,
-                name: name,
-                valueType,
-                mutability,
-                initExpr);
-            bool added = _definedGlobals.TryAdd(name, global);
-            Debug.Assert(added, $"Duplicate global name: {name}");
-
-            WasmGlobalSection section = GetOrCreateSection<WasmGlobalSection>(
-                WasmObjectNodeSection.GlobalSection,
-                out SectionWriter writer);
-            section.WriteEntry(writer, global);
-        }
-
-
-        private void WriteGlobalSection()
+        private protected override void WriteGlobalSection()
         {
             // webcilVersion: i32 const = 0
             WriteGlobal("webcilVersion", WasmValueType.I32, WasmMutabilityType.Const,
                 new WasmInstructionGroup([new WasmConstExpr(WasmExprKind.I32Const, WebcilConstants.WC_VERSION_MAJOR)]));
-        }
-
-        // Sections excluding Webcil Data segment
-        readonly string[] SectionOrder =
-        [
-            ObjectNodeSection.WasmTypeSection.Name,
-            WasmObjectNodeSection.ImportSection.Name,
-            WasmObjectNodeSection.FunctionSection.Name,
-            WasmObjectNodeSection.GlobalSection.Name,
-            WasmObjectNodeSection.ExportSection.Name,
-            WasmObjectNodeSection.ElementSection.Name,
-            WasmObjectNodeSection.DataCountSection.Name,
-            ObjectNodeSection.WasmCodeSection.Name,
-        ];
-
-        private int[] _sectionEmitOrder = null;
-        private int[] SectionEmitOrder
-        {
-            get
-            {
-                if (_sectionEmitOrder == null)
-                {
-                    _sectionEmitOrder = SectionOrder
-                        .Where(name => _sections.Contains(name))
-                        .Select(name => _sections.GetSectionIndex(name))
-                        .ToArray();
-                }
-
-                return _sectionEmitOrder;
-            }
         }
 
         private static readonly ObjectNodeSection WebcilRelocSection = new ObjectNodeSection("reloc", SectionType.ReadOnly);
@@ -650,7 +342,7 @@ namespace ILCompiler.ObjectWriter
 
             if (_pendingBaseRelocs.Count > 0)
             {
-                GetOrCreateSection<WebcilSection>(WebcilRelocSection);
+                GetOrCreateSection<WebcilSection>(WebcilRelocSection, out _);
             }
 
             WebcilSection[] webcilSections = _sections.Sections.OfType<WebcilSection>().ToArray();
@@ -796,25 +488,6 @@ namespace ILCompiler.ObjectWriter
                 }
             }
         }
-
-        private readonly WasmSymbolManager _wasmSymbolManager = new();
-
-        /// <summary>
-        /// Maps an Import kind to the WasmIndexSpace that it can be referenced within.
-        /// Imports will always be the first logical entries in their respective index spaces.
-        /// </summary>
-        private static WasmIndexSpace GetIndexSpace(WasmExternalKind kind) => kind switch
-        {
-            WasmExternalKind.Function => WasmIndexSpace.Function,
-            WasmExternalKind.Table => WasmIndexSpace.Table,
-            WasmExternalKind.Memory => WasmIndexSpace.Memory,
-            WasmExternalKind.Global => WasmIndexSpace.Global,
-            WasmExternalKind.Tag => WasmIndexSpace.Tag,
-            _ => throw new ArgumentOutOfRangeException(nameof(kind)),
-        };
-
-        private void RegisterFunctionSymbol(Utf8String name) =>
-            _wasmSymbolManager.AddDefinition(name, WasmIndexSpace.Function);
 
         /// <summary>
         /// Processes the deferred file-level relocations after webcil section VirtualAddresses
@@ -1219,16 +892,17 @@ namespace ILCompiler.ObjectWriter
         }
 #nullable disable
 
-        public const int StackPointerGlobalIndex = WasmGlobalImports.StackPointerGlobalIndex;
-        public const int ImageBaseGlobalIndex = WasmGlobalImports.ImageBaseGlobalIndex;
-        public const int TableBaseGlobalIndex = WasmGlobalImports.TableBaseGlobalIndex;
-        public const int AsyncContinuationGlobalIndex = WasmGlobalImports.AsyncContinuationGlobalIndex;
         public const int RtlRestoreContextTagIndex = 0;
         private static readonly Utf8String RtlRestoreContextTagName = new("rtlRestoreContextTag");
 
         private static readonly WasmFuncType RtlRestoreContextTagSignature = new(
             new([]),
             new([]));
+
+        internal const int StackPointerGlobalIndex = 0;
+        internal const int ImageBaseGlobalIndex = 1;
+        internal const int TableBaseGlobalIndex = 2;
+        internal const int AsyncContinuationGlobalIndex = 3;
 
         private WasmImport[] CreateDefaultGlobalImports()
         {
@@ -1245,25 +919,7 @@ namespace ILCompiler.ObjectWriter
             ];
         }
 
-        private int RegisterSignature(WasmFuncType signature)
-        {
-            Utf8String signatureKey = signature.GetMangledName(_nodeFactory.NameMangler);
-            if (_wasmSymbolManager.TryGetSymbol(signatureKey, out WasmSymbol signatureSymbol))
-            {
-                return signatureSymbol.Index;
-            }
-
-            GetOrCreateSection(ObjectNodeSection.WasmTypeSection, out SectionWriter typeSectionWriter);
-            byte[] encodedSignature = new byte[signature.EncodeSize()];
-            signature.Encode(encodedSignature);
-            _wasmSymbolManager.AddDefinition(signatureKey, WasmIndexSpace.Type);
-            typeSectionWriter.EmitSymbolDefinition(signatureKey);
-            typeSectionWriter.EmitData(encodedSignature);
-
-            return _wasmSymbolManager.GetSymbol(signatureKey).Index;
-        }
-
-        private void WriteImports()
+        private protected override void WriteImports()
         {
             foreach (WasmImport import in CreateDefaultGlobalImports())
             {
@@ -1280,7 +936,7 @@ namespace ILCompiler.ObjectWriter
             WriteImport(memoryImport);
         }
 
-        private void WriteExports()
+        private protected override void WriteExports()
         {
             WriteTableExport("table", 0);
 
@@ -1297,13 +953,7 @@ namespace ILCompiler.ObjectWriter
             }
         }
 
-        /// <summary>
-        /// This dictionary contains the map of symbol names to their location in the object file.
-        /// The SymbolDefinitions returned do not encode the logical index used in the Wasm module, and should not be used for
-        /// resolving relocations for INDEX relocation kinds.
-        /// </summary>
-        private Dictionary<Utf8String, SymbolDefinition> _definedSymbols;
-        private void WriteElements()
+        private protected override void WriteElements()
         {
             // Generate the function pointer table element that contains function pointers for all of our functions.
             // Function indices are assigned sequentially (0..MethodCount-1) so that
@@ -1313,30 +963,6 @@ namespace ILCompiler.ObjectWriter
                 .ToArray();
 
             WriteElementSegment(functionIndices);
-        }
-
-        // Populate the sections whose entries are derived from the completed symbol table.
-        private protected override void EmitSymbolTable(IDictionary<Utf8String, SymbolDefinition> definedSymbols, SortedSet<Utf8String> undefinedSymbols)
-        {
-            WriteImports();
-            WriteGlobalSection();
-            WriteExports();
-            WriteElements();
-
-            // Register defined symbols for future use during relocation resolution
-            _definedSymbols = new Dictionary<Utf8String, SymbolDefinition>(definedSymbols);
-        }
-
-        private void FinalizeSectionEntryCounts()
-        {
-            _sections.GetSection<WasmExternallyCountedSection>(ObjectNodeSection.WasmTypeSection.Name)
-                .SetEntryCount(_wasmSymbolManager.GetDefinitionCount(WasmIndexSpace.Type));
-            _sections.GetSection<WasmExternallyCountedSection>(ObjectNodeSection.WasmCodeSection.Name)
-                .SetEntryCount(MethodCount);
-
-            Debug.Assert(_sections.GetSection<WasmFunctionSection>(WasmObjectNodeSection.FunctionSection.Name).EntryCount == MethodCount);
-            Debug.Assert(_sections.GetSection<WasmImportSection>(WasmObjectNodeSection.ImportSection.Name).EntryCount == _wasmSymbolManager.GetImportCount());
-            Debug.Assert(_sections.GetSection<WasmGlobalSection>(WasmObjectNodeSection.GlobalSection.Name).EntryCount == _wasmSymbolManager.GetDefinitionCount(WasmIndexSpace.Global));
         }
     }
 }
